+++ lighttpd/tox.ini
@@ -8,6 +8,8 @@ envlist =
 [testenv]
 usedevelop = true
 platform = linux|darwin|win32
+skip_install =
+    flake8: true
 passenv =
     DOCKER*
     COMPOSE*
@@ -16,7 +18,7 @@ deps =
     -rrequirements-dev.txt
 commands =
     pip install -r requirements.in
-    pytest -v
+    pytest -m"integration" -v
 
 [testenv:flake8]
 skip_install = true
@@ -24,5 +26,5 @@ deps = flake8
 commands = flake8 .
 
 [flake8]
-exclude = .eggs,.tox
+exclude = .eggs,.tox,build
 max-line-length = 120
